// src/renderer/src/types/language.ts
export type Language = 'de_DE' | 'en_CA' | 'en_US' | 'es_ES' | 'es_MX' | 'fr_CA' | 'fr_FR' | 'it_IT' | 'pt_PT' | 'pt_BR';

export type LanguageJson = {
    global_app_name: string;  
    global_form_input_error_required: string; 
    global_back_link: string; 
    global_form_message_error_title: string; 
    global_form_message_error_body: string; 
    global_form_message_success_tittle: string; 
    global_form_message_success_body: string; 
    sigin_title: string; 
    sigin_form_label_email: string; 
    sigin_form_label_password: string; 
    sigin_form_label_submit: string; 
    main_title: string; 
    main_subtitle: string; 
    welcome_title: string; 
    welcome_subtitle: string; 
    welcome_form_label_register: string; 
    welcome_form_label_sigin: string; 
    register_title: string; 
    register_form_label_fn: string; 
    register_form_label_ln: string; 
    register_form_label_email: string; 
    register_form_label_password: string; 
    register_form_label_password_2: string; 
    register_form_label_submit: string; 
  }